refactor #192 : 빌더 패턴 클래스, 싱글톤 인스턴스 클래스 파일 분리
